+++ src/rajawali/animation/RotateAnimation3D.java
@@ -1,3 +1,15 @@
+/**
+ * Copyright 2013 Dennis Ippel
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package rajawali.animation;
 
 import rajawali.ATransformable3D;
@@ -39,8 +51,7 @@ public class RotateAnimation3D extends Animation3D {
 		mAngleAxisRotation = true;
 		mQuat = new Quaternion();
 		mQuatFrom = new Quaternion();
-		// TODO: Switch Quaternions to take in doubles instead of floats.
-		mQuatFrom.fromAngleAxis(axis, (float) rotateFrom);
+		mQuatFrom.fromAngleAxis(axis, rotateFrom);
 		mRotationAxis = axis;
 		mRotateFrom = rotateFrom;
 		mDegreesToRotate = degreesToRotate;
@@ -58,9 +69,9 @@ public class RotateAnimation3D extends Animation3D {
 		mRotateY = yRotate;
 		mRotateZ = zRotate;
 
-		mQuat.multiply(new Quaternion().fromAngleAxis(Vector3.getAxisVector(Axis.Y), (float) yRotate));
-		mQuat.multiply(new Quaternion().fromAngleAxis(Vector3.getAxisVector(Axis.Z), (float) zRotate));
-		mQuat.multiply(new Quaternion().fromAngleAxis(Vector3.getAxisVector(Axis.X), (float) xRotate));
+		mQuat.multiply(new Quaternion().fromAngleAxis(Vector3.getAxisVector(Axis.Y), yRotate));
+		mQuat.multiply(new Quaternion().fromAngleAxis(Vector3.getAxisVector(Axis.Z), zRotate));
+		mQuat.multiply(new Quaternion().fromAngleAxis(Vector3.getAxisVector(Axis.X), xRotate));
 	}
 
 	public RotateAnimation3D(Vector3 rotate) {
@@ -86,11 +97,11 @@ public class RotateAnimation3D extends Animation3D {
 		if (mAngleAxisRotation) {
 			// Rotation around an axis by amount of degrees.
 			mRotationAngle = mRotateFrom + (mInterpolatedTime * mDegreesToRotate);
-			mQuat.fromAngleAxis(mRotationAxis, (float) mRotationAngle);
+			mQuat.fromAngleAxis(mRotationAxis, mRotationAngle);
 			mQuat.multiply(mQuatFrom);
 			mTransformable3D.setOrientation(mQuat);
 		} else {
-			mTransformable3D.setOrientation(Quaternion.slerpAndCreate(mQuatFrom, mQuat, (float)mInterpolatedTime));
+			mTransformable3D.setOrientation(Quaternion.slerpAndCreate(mQuatFrom, mQuat, mInterpolatedTime));
 		}
 	}
 }
