+++ src/rajawali/parser/Max3DSParser.java
@@ -1,3 +1,15 @@
+/**
+ * Copyright 2013 Dennis Ippel
+ * 
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * 
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
 package rajawali.parser;
 
 import java.io.BufferedInputStream;
@@ -193,58 +205,58 @@ public class Max3DSParser extends AMeshParser {
 
 				coord = vertices.get(v1);
 
-				aVertices[ic++] = coord.x;
-				aVertices[ic++] = coord.y;
-				aVertices[ic++] = coord.z;
+				aVertices[ic++] = (float) coord.x;
+				aVertices[ic++] = (float) coord.y;
+				aVertices[ic++] = (float) coord.z;
 
 				aIndices[ivi] = ivi++;
 
 				coord = vertices.get(v2);
-				aVertices[ic++] = coord.x;
-				aVertices[ic++] = coord.y;
-				aVertices[ic++] = coord.z;
+				aVertices[ic++] = (float) coord.x;
+				aVertices[ic++] = (float) coord.y;
+				aVertices[ic++] = (float) coord.z;
 
 				aIndices[ivi] = ivi++;
 
 				coord = vertices.get(v3);
-				aVertices[ic++] = coord.x;
-				aVertices[ic++] = coord.y;
-				aVertices[ic++] = coord.z;
+				aVertices[ic++] = (float) coord.x;
+				aVertices[ic++] = (float) coord.y;
+				aVertices[ic++] = (float) coord.z;
 
 				aIndices[ivi] = ivi++;
 
 				if (texCoords != null && texCoords.size() > 0) {
 					texcoord = texCoords.get(v1);
 
-					aTexCoords[itc++] = texcoord.x;
-					aTexCoords[itc++] = texcoord.y;
+					aTexCoords[itc++] = (float) texcoord.x;
+					aTexCoords[itc++] = (float) texcoord.y;
 
 					texcoord = texCoords.get(v2);
 
-					aTexCoords[itc++] = texcoord.x;
-					aTexCoords[itc++] = texcoord.y;
+					aTexCoords[itc++] = (float) texcoord.x;
+					aTexCoords[itc++] = (float) texcoord.y;
 
 					texcoord = texCoords.get(v3);
 
-					aTexCoords[itc++] = texcoord.x;
-					aTexCoords[itc++] = texcoord.y;
+					aTexCoords[itc++] = (float) texcoord.x;
+					aTexCoords[itc++] = (float) texcoord.y;
 				}
 
 				normal = vertNormals.get(v1);
-				aNormals[itn++] = normal.x;
-				aNormals[itn++] = normal.y;
-				aNormals[itn++] = normal.z;
+				aNormals[itn++] = (float) normal.x;
+				aNormals[itn++] = (float) normal.y;
+				aNormals[itn++] = (float) normal.z;
 				normal = vertNormals.get(v2);
 
-				aNormals[itn++] = normal.x;
-				aNormals[itn++] = normal.y;
-				aNormals[itn++] = normal.z;
+				aNormals[itn++] = (float) normal.x;
+				aNormals[itn++] = (float) normal.y;
+				aNormals[itn++] = (float) normal.z;
 
 				normal = vertNormals.get(v3);
 
-				aNormals[itn++] = normal.x;
-				aNormals[itn++] = normal.y;
-				aNormals[itn++] = normal.z;
+				aNormals[itn++] = (float) normal.x;
+				aNormals[itn++] = (float) normal.y;
+				aNormals[itn++] = (float) normal.z;
 			}
 
 			BaseObject3D targetObj = new BaseObject3D(mObjNames.get(j));
